backend/requirements.txt
"fastapi[standard]"
mysql-connector-python 
python-dotenv